contratos/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-swagger>=0.2.14
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
pulsar-client==3.7.0
pulsar-client[avro]==3.7.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.25.0
PyDispatcher==2.0.6